[Refactor] create Redirect folder(#61)
